% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index_number.R
\name{generar_indice}
\alias{generar_indice}
\title{función para calcular números índice}
\usage{
generar_indice(serie, fecha, fecha_base)
}
\arguments{
\item{serie}{vector con la serie}

\item{fecha}{vector con fechas}

\item{fecha_base}{período base.}
}
\description{
función para calcular números índice
}
