% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_tablas_bea.R
\name{leer_tablas_bea}
\alias{leer_tablas_bea}
\title{Función para leer las tablas de bea}
\usage{
leer_tablas_bea(path)
}
\arguments{
\item{path}{el nombre del directorio donde se encuentra la tabla}
}
\description{
Función para leer las tablas de bea
}
